Add validation when updating the product image

+++ app/Actions/Products/UpdateProduct.php
@@ -34,7 +34,9 @@ class UpdateProduct
      */
     public function uploadFile(Request $request, Product $product): string
     {
-        Storage::disk('products')->delete($product->image);
+        if ($product->image) {
+            Storage::disk('products')->delete($product->image);
+        }
 
         $upload = $request->file('image');
         $filename = Str::uuid() . '.' . $upload->extension();
